table and form work

+++ db/migrate/20211129150043_create_interventions.rb
@@ -1,12 +1,12 @@
 class CreateInterventions < ActiveRecord::Migration[5.2]
   def change
     create_table :interventions do |t|
-      t.integer :author
-      t.integer :customer_id
-      t.integer :building_id
-      t.integer :battery_id, null: true
-      t.integer :column_id, null: true
-      t.integer :elevator_id, null: true
+      t.bigint :author
+      t.bigint :customer_id
+      t.bigint :building_id
+      t.bigint :battery_id, null: true
+      t.bigint :column_id, null: true
+      t.bigint :elevator_id, null: true
       t.integer :employee_id, null: true
       t.datetime :start_of_intervention, null: true
       t.datetime :end_of_intervention, null: true
